// src/app/shared/models/address-info.model.ts
export class AddressInfo {
  locationName: string = '';
  locationId: string = '';
  addressLine1: string = '';
  addressLine2: string = '';
  addressLine3: string = '';
  city: string = '';
  state: string = '';
  postalCode: string = '';
  countryCode: string = '';
  phoneNumber: string = '';
  contactName: string = '';
  customerName: string = '';
  taxId: string = '';
}